test(cloudflare-metrics): split monolithic test file into per-module files
the 1350-line index.test.ts covered seven unrelated subjects and made it
hard to spot what was and wasn't tested. split it module-by-module, move
the shared recording provider / fake graphql & rest clients into a new
test-helpers.ts, and add two small coverage gaps that surfaced during
the split:

- graphql client: verify scheduled invocations are only attached to the
  first chunked fetch (the subtle correctness guard in the chunking fix)
- graphql client: verify requestCount / errorResponseCount gauges are
  incremented correctly across multiple chunks

new layout:
  cloudflare-api.test.ts    # rest client pagination, errors, getZone
  collector.test.ts         # row emission, enrichment, zone iteration
  datasets.test.ts          # dataset registry invariants
  graphql-client.test.ts    # query builders, chunking, counters
  handler.test.ts           # SELF.fetch http handler tests
  metrics.test.ts           # Metric class + InfluxMetricsProvider
  test-helpers.ts           # shared fakes (RecordingProvider, FakeGraphQLClient...)
